move to util file
