+++ src/utils/datasets.ts
@@ -1,8 +1,8 @@
-import { GalileoApiClient, DatasetRow } from '../api-client';
+import { DatasetRow, GalileoApiClient } from '../api-client';
 import {
-  createDatasetRecord,
   Dataset,
-  DatasetRecord
+  DatasetRecord,
+  DatasetRecordOptions
 } from '../types/dataset.types';
 import { existsSync, PathLike, writeFileSync } from 'fs';
 import { tmpdir } from 'os';
@@ -195,6 +195,68 @@ export const getDatasetContent = async ({
   }
 
   return await apiClient.getDatasetContent(datasetId!);
+};
+
+export const createDatasetRecord = ({
+  id,
+  input,
+  output,
+  metadata
+}: DatasetRecordOptions): DatasetRecord => {
+  let resultMetadata: Record<string, string> | undefined = undefined;
+  if (metadata != null) {
+    // checks null & undefined
+    let record: Record<string, unknown> = {};
+    if (typeof metadata === 'string') {
+      try {
+        record = JSON.parse(metadata);
+      } catch (error) {
+        if (
+          error instanceof SyntaxError &&
+          error.message.includes('JSON.parse')
+        ) {
+          record = { metadata: metadata };
+        } else {
+          throw error;
+        }
+      }
+    } else if (typeof metadata === 'object') {
+      record = metadata as Record<string, unknown>;
+    } else {
+      throw new Error('Dataset metadata must be a string or object');
+    }
+    for (const value of Object.values(record)) {
+      if (typeof value !== 'string') {
+        throw new Error('Dataset metadata values must be strings');
+      }
+    }
+    resultMetadata = record as Record<string, string>;
+  }
+
+  return {
+    id,
+    input: serializeToString(input),
+    output: output === undefined ? undefined : serializeToString(output),
+    metadata: resultMetadata
+  };
+};
+
+const serializeToString = (value: unknown): string => {
+  return typeof value === 'string' ? value : JSON.stringify(value);
+};
+
+export const deserializeInputFromString = (
+  value: string
+): Record<string, unknown> => {
+  try {
+    return JSON.parse(value);
+  } catch (error) {
+    if (error instanceof SyntaxError && error.message.includes('JSON.parse')) {
+      return { value: value };
+    } else {
+      throw error;
+    }
+  }
 };
 
 export const getRecordsForDataset = async ({
